+++ src/lib/components/List/ListItemIcon.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+interface IListItemIconProps {
+	icon: React.SVGProps<SVGSVGElement>;
+}
+
+export const ListItemIcon = ({ icon }: IListItemIconProps) => (
+	<div className="icon">
+		<div className="svg-icon">
+			{icon}
+		</div>
+	</div>
+)