+++ src/showcase/assets/index.ts
@@ -1,2 +1,3 @@
 export * from './InfoSolidIcon';
 export * from './BarsSolidIcon';
+export * from './UserCircleSolidIcon';