+++ src/showcase/examples/ListPage.tsx
@@ -1,18 +1,9 @@
 import React, { Fragment } from 'react';
-import {
-	List, ListItem, ListItemAvatar, HomeSolidIcon, ListItemAction, IconButton, DropDown, DropDownItem, ListItemText, snackbarService
-} from '../../lib';
+import { List, ListItem, ListItemAvatar, HomeSolidIcon, ListItemAction, IconButton, ListItemText, ListItemIcon } from '../../lib';
+import { UserCircleSolidIcon } from '../assets';
 import { withOptions } from './components';
 
 export const ListPageBase = () => {
-
-	const handleClickItem = () => {
-		snackbarService.show('Item clicked');
-	}
-
-	const handleClickDropDownItem = () => {
-		snackbarService.show('DropdownItem clicked');
-	}
 
 	return (
 		<Fragment>
@@ -25,7 +16,7 @@ export const ListPageBase = () => {
 				<ListItem><ListItemText primary="lorem ipsum" /></ListItem>
 			</List>
 
-			<h3>Two lines</h3>
+			<h3 className="mt-3">Two lines</h3>
 			<List>
 				<ListItem><ListItemText primary="1st lorem ipsum" secondary="2nd lorem ipsum" /></ListItem>
 				<ListItem><ListItemText primary="1st lorem ipsum" secondary="2nd lorem ipsum" /></ListItem>
@@ -33,25 +24,32 @@ export const ListPageBase = () => {
 				<ListItem><ListItemText primary="1st lorem ipsum" secondary="2nd lorem ipsum" /></ListItem>
 			</List>
 
-			<h3 className="mt-3">extended</h3>
+			<h3 className="mt-3">Avatar</h3>
 			<List>
-				<ListItem onClick={handleClickItem}>
-					<ListItemAvatar avatar={<HomeSolidIcon />} />
-
+				<ListItem>
+					<ListItemAvatar avatar={<UserCircleSolidIcon />} />
 					<ListItemText primary="lorem ipsum" />
+				</ListItem>
+			</List>
 
+			<h3 className="mt-3">Icon</h3>
+			<List>
+				<ListItem>
+					<ListItemIcon icon={<HomeSolidIcon />} />
+					<ListItemText primary="lorem ipsum" />
+				</ListItem>
+			</List>
+
+			<h3 className="mt-3">Action Item</h3>
+			<List>
+				<ListItem>
+					<ListItemText primary="lorem ipsum" />
 					<ListItemAction>
-						<DropDown menuPosition="right" toggle={<IconButton icon={<HomeSolidIcon />} />}>
-							<DropDownItem onClick={handleClickDropDownItem}>
-								test 1
-							</DropDownItem>
-							<DropDownItem onClick={handleClickDropDownItem}>
-								test 2
-							</DropDownItem>
-						</DropDown>
+						<IconButton icon={<HomeSolidIcon />} />
 					</ListItemAction>
 				</ListItem>
 			</List>
+
 		</Fragment>
 	);
 }